--- roshambo/pom.xml
@@ -1,22 +1,0 @@
-<project>
-	<modelVersion>4.0.0</modelVersion>
-	<!-- interesting -->
-	<groupId>com.mcnz.rps</groupId>
-	<artifactId>roshambo</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<packaging>jar</packaging>
-
-	<name>roshambo</name>
-
-
-	<dependencies>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.11</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-
-</project>
